Island: Relocate aws_service from common to monkey_island
