+++ monkey/monkey_island/cc/services/aws_service.py
