+++ monkey/monkey_island/cc/services/reporting/aws_exporter.py
@@ -5,8 +5,8 @@ from datetime import datetime
 import boto3
 from botocore.exceptions import UnknownServiceError
 
-from common.aws import aws_service
 from common.aws.aws_instance import AwsInstance
+from monkey_island.cc.services import aws_service
 from monkey_island.cc.services.reporting.exporter import Exporter
 
 __authors__ = ["maor.rayzin", "shay.nehmad"]
